# template/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "FirstWeb项目API接口说明，支持用户认证和数据库通用操作",
        "title": "FirstWeb API接口文档",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@firstweb.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "description": "用户登录接口，使用SM2解密密码并验证用户身份",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户认证"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录参数",
                        "name": "loginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sysbase.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "101": {
                        "description": "解密失败或其他错误",
                        "schema": {
                            "$ref": "#/definitions/sysbase.ErrorResponse"
                        }
                    },
                    "102": {
                        "description": "无效的请求参数",
                        "schema": {
                            "$ref": "#/definitions/sysbase.ErrorResponse"
                        }
                    },
                    "103": {
                        "description": "用户名或密码错误",
                        "schema": {
                            "$ref": "#/definitions/sysbase.ErrorResponse"
                        }
                    },
                    "200": {
                        "description": "登录成功，返回token",
                        "schema": {
                            "$ref": "#/definitions/sysbase.LoginResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "sysbase.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "type": "integer",
                    "example": 101
                },
                "data": {
                    "description": "错误数据",
                    "type": "string",
                    "example": "null"
                },
                "msg": {
                    "description": "错误消息",
                    "type": "string",
                    "example": "错误信息"
                }
            }
        },
        "sysbase.LoginRequest": {
            "type": "object",
            "required": [
                "code",
                "password"
            ],
            "properties": {
                "code": {
                    "description": "用户名",
                    "type": "string",
                    "example": "admin"
                },
                "password": {
                    "description": "密码(SM2加密)",
                    "type": "string",
                    "example": "123456"
                }
            }
        },
        "sysbase.LoginResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "返回的token",
                    "type": "string",
                    "example": "token-uuid"
                },
                "msg": {
                    "description": "响应消息",
                    "type": "string",
                    "example": "操作成功"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "TokenAuth": {
            "description": "Token认证",
            "type": "apiKey",
            "name": "Token",
            "in": "header"
        }
    }
}